+++ DemoApps/GRDBDemoiOS/GRDBDemoiOS/PersonsViewController.swift
@@ -47,7 +47,7 @@ class PersonsViewController: UITableViewController {
             recordsDidChange: { [unowned self] _ in
                 self.tableView.endUpdates()
             })
-        personsController.performFetch()
+        try! personsController.performFetch()
         
         configureToolbar()
     }
@@ -176,7 +176,7 @@ extension PersonsViewController {
         setEditing(false, animated: true)
         
         try! dbQueue.inTransaction { db in
-            for person in Person.fetch(db) {
+            for person in try Person.fetchAll(db) {
                 person.score = Person.randomScore()
                 try person.update(db)
             }
@@ -190,7 +190,7 @@ extension PersonsViewController {
         for _ in 0..<50 {
             DispatchQueue.global().async {
                 try! dbQueue.inTransaction { db in
-                    if Person.fetchCount(db) == 0 {
+                    if try Person.fetchCount(db) == 0 {
                         // Insert persons
                         for _ in 0..<8 {
                             try Person(name: Person.randomName(), score: Person.randomScore()).insert(db)
@@ -203,12 +203,12 @@ extension PersonsViewController {
                         }
                         // Delete a person
                         if arc4random_uniform(2) == 0 {
-                            if let person = Person.order(sql: "RANDOM()").fetchOne(db) {
+                            if let person = try Person.order(sql: "RANDOM()").fetchOne(db) {
                                 try person.delete(db)
                             }
                         }
                         // Update some persons
-                        for person in Person.fetchAll(db) {
+                        for person in try Person.fetchAll(db) {
                             if arc4random_uniform(2) == 0 {
                                 person.score = Person.randomScore()
                                 try person.update(db)